requirements.txt
fastapi
uvicorn[standard]
ultralytics
python-multipart
torch>=1.13.0
torchvision
efficientnet_pytorch
pillow
joblib
scikit-learn